# obl/oblas_lite: add OBLAS_TINY to disable look-up table